Tests
Some corrections in the examples and implementation of the tests
