+++ biorbd_examples/two_pendulums.py
@@ -1,35 +1,44 @@
 """
 This script is used to integrate the motion with a variational integrator based on the discrete Lagrangian,
 and a first order quadrature method.
-This example is a simple pendulum freed and constrained in translation on the y-axis and the z-axis.
+This example is a double pendulum but the second one has no parent, it is constrained to be coincident with the first
+one.
 """
 import biorbd_casadi
 
 from casadi import MX, jacobian, Function
 
-from varint.minimal_variational_integrator import VariationalIntegrator, QuadratureRule
+from varint.minimal_variational_integrator import VariationalIntegrator
 
 from utils import *
 
 
-def one_pendulum(time: float = 10, time_step: float = 0.05, unit_test: bool = False):
-    biorbd_casadi_model = biorbd_casadi.Model(Models.ONE_PENDULUM.value)
-    biorbd_model = biorbd.Model(Models.ONE_PENDULUM.value)
+def two_pendulums(time: float = 10, time_step: float = 0.05, unit_test: bool = False):
+    biorbd_casadi_model = biorbd_casadi.Model(Models.TWO_PENDULUMS.value)
+    biorbd_model = biorbd.Model(Models.TWO_PENDULUMS.value)
 
     import time as t
 
     tic0 = t.time()
 
-    q_t0 = np.array([1.54])
-    q_t1 = np.array([1.545])
+    # Rotations at t0
+    q_t0 = np.array([1.54, 1.54])
+    # Translations between Seg0 and Seg1 at t0, calculated with cos and sin as Seg1 has no parent
+    t_t0 = np.array([np.sin(q_t0[0]), -np.cos(q_t0[0])])
+    # Rotations at t1
+    q_t1 = np.array([1.545, 1.545])
+    # Translations between Seg0 and Seg1 at t1, calculated with cos and sin as Seg1 has no parent
+    t_t1 = np.array([np.sin(q_t1[0]), -np.cos(q_t1[0])])
 
-    all_q_t0 = np.array([0, 0, q_t0[0]])
-    all_q_t1 = np.array([0, 0, q_t1[0]])
+    all_q_t0 = np.array([q_t0[0], t_t0[0], t_t0[1], q_t0[1]])
+    all_q_t1 = np.array([q_t1[0], t_t1[0], t_t1[1], q_t1[1]])
 
     # build  constraint
     # the origin of the second pendulum is constrained to the tip of the first pendulum
     q_sym = MX.sym("q", (biorbd_casadi_model.nbQ(), 1))
-    constraint = biorbd_casadi_model.markers(q_sym)[0].to_mx()[1:] - MX.zeros((2, 1))
+    constraint = (
+        biorbd_casadi_model.markers(q_sym)[1].to_mx()[1:] - biorbd_casadi_model.globalJCS(q_sym, 1).to_mx()[1:3, 3]
+    )
     fcn_constraint = Function("constraint", [q_sym], [constraint], ["q"], ["constraint"]).expand()
     fcn_jacobian = Function("jacobian", [q_sym], [jacobian(constraint, q_sym)], ["q"], ["jacobian"]).expand()
 
@@ -43,10 +52,8 @@ def one_pendulum(time: float = 10, time_step: float = 0.05, unit_test: bool = Fa
         time=time,
         constraints=fcn_constraint,
         jac=fcn_jacobian,
-        discrete_approximation=QuadratureRule.TRAPEZOIDAL,
         q_init=np.concatenate((all_q_t0[:, np.newaxis], all_q_t1[:, np.newaxis]), axis=1),
     )
-    # vi.set_initial_values(q_prev=1.54, q_cur=1.545)
     q_vi, lambdas_vi = vi.integrate()
 
     tic2 = t.time()
@@ -55,7 +62,7 @@ def one_pendulum(time: float = 10, time_step: float = 0.05, unit_test: bool = Fa
     if unit_test:
         import bioviz
 
-        b = bioviz.Viz(Models.ONE_PENDULUM.value)
+        b = bioviz.Viz(Models.TWO_PENDULUMS.value)
         b.load_movement(q_vi)
         b.exec()
 
@@ -72,22 +79,28 @@ def one_pendulum(time: float = 10, time_step: float = 0.05, unit_test: bool = Fa
             np.arange(0, time, time_step), q_vi[2, :], label="Variational Integrator", color="red", linestyle="-"
         )
         axs[0, 1].plot(
-            np.arange(0, time, time_step), lambdas_vi[0, :], label="Variational Integrator", color="red", linestyle="-"
+            np.arange(0, time, time_step), q_vi[3, :], label="Variational Integrator", color="red", linestyle="-"
         )
         axs[1, 1].plot(
+            np.arange(0, time, time_step), lambdas_vi[0, :], label="Variational Integrator", color="red", linestyle="-"
+        )
+        axs[2, 1].plot(
             np.arange(0, time, time_step), lambdas_vi[1, :], label="Variational Integrator", color="red", linestyle="-"
         )
 
         axs[0, 0].set_title("q0")
         axs[1, 0].set_title("q1")
         axs[2, 0].set_title("q2")
-        axs[0, 1].set_title("lambda1")
-        axs[1, 1].set_title("lambda2")
+        axs[0, 1].set_title("q3")
+        axs[1, 1].set_title("lambda0")
+        axs[2, 1].set_title("lambda1")
+        axs[0, 0].legend()
 
-        # plot total energy for both methods
+        # Plot total energy
         plt.figure()
         plt.plot(discrete_total_energy(biorbd_model, q_vi, time_step), label="Variational Integrator", color="red")
-        plt.title("Total energy")
+        plt.title("Total energy with variational integrator")
+        plt.legend()
 
         # verify the constraint respect
         plt.figure()
@@ -102,4 +115,4 @@ def one_pendulum(time: float = 10, time_step: float = 0.05, unit_test: bool = Fa
 
 
 if __name__ == "__main__":
-    one_pendulum(unit_test=True)
+    two_pendulums(unit_test=True)
